DotNetTraining

+++ DoubleCalculator/Program.cs
@@ -1,12 +1,12 @@
 ﻿
-int i, j, res = 0;
+double i, j, res = 0;
 string oper;
 
 Console.Write("Value 1 : ");
-i = Int32.Parse(Console.ReadLine());
+i = double.Parse(Console.ReadLine());
 
 Console.Write("Value 2 : ");
-j = Int32.Parse(Console.ReadLine());
+j = double.Parse(Console.ReadLine());
 
 Console.Write("Enter operator (+, -,*,/) : ");
 oper = Console.ReadLine();
